Remove ajax_select from counters
